+++ bin/faucet/src/api/mod.rs
@@ -7,13 +7,13 @@ use axum::Router;
 use axum::extract::FromRef;
 use axum::routing::get;
 use http::{HeaderValue, Request};
-use miden_client::account::AccountId;
+use miden_client::account::{AccountId, AccountIdError, AddressError};
 use miden_client::store::Store;
 use miden_client::utils::RwLock;
 use miden_faucet_lib::FaucetId;
 use miden_faucet_lib::requests::MintRequestSender;
 use miden_faucet_lib::types::AssetAmount;
-use miden_faucet_pow::{ApiKey, PoW, PoWConfig};
+use miden_pow_rate_limiter::{PoWRateLimiter, PoWRateLimiterConfig};
 use sha3::{Digest, Sha3_256};
 use tokio::net::TcpListener;
 use tower::ServiceBuilder;
@@ -28,6 +28,7 @@ use crate::api::get_metadata::{Metadata, get_metadata};
 use crate::api::get_note::get_note;
 use crate::api::get_pow::get_pow;
 use crate::api::get_tokens::{GetTokensState, MintRequestError, get_tokens};
+use crate::api_key::ApiKey;
 
 mod frontend;
 mod get_metadata;
@@ -43,7 +44,7 @@ mod get_tokens;
 pub struct Server {
     mint_state: GetTokensState,
     metadata: &'static Metadata,
-    pow: PoW,
+    rate_limiter: PoWRateLimiter,
     api_keys: HashSet<ApiKey>,
     store: Arc<dyn Store>,
 }
@@ -58,9 +59,10 @@ impl Server {
         max_claimable_amount: AssetAmount,
         mint_request_sender: MintRequestSender,
         pow_secret: &str,
-        pow_config: PoWConfig,
+        rate_limiter_config: PoWRateLimiterConfig,
         api_keys: &[ApiKey],
         store: Arc<dyn Store>,
+        explorer_url: Option<Url>,
     ) -> Self {
         let mint_state = GetTokensState::new(mint_request_sender, max_claimable_amount);
         let metadata = Metadata {
@@ -68,6 +70,7 @@ impl Server {
             issuance,
             max_supply,
             decimals,
+            explorer_url,
         };
         // SAFETY: Leaking is okay because we want it to live as long as the application.
         let metadata = Box::leak(Box::new(metadata));
@@ -77,12 +80,12 @@ impl Server {
         hasher.update(pow_secret.as_bytes());
         let secret_bytes: [u8; 32] = hasher.finalize().into();
 
-        let pow = PoW::new(secret_bytes, pow_config);
+        let rate_limiter = PoWRateLimiter::new(secret_bytes, rate_limiter_config);
 
         Server {
             mint_state,
             metadata,
-            pow,
+            rate_limiter,
             api_keys: api_keys.iter().cloned().collect::<HashSet<_>>(),
             store,
         }
@@ -170,14 +173,17 @@ impl Server {
         challenge: &str,
         nonce: u64,
         account_id: AccountId,
-        api_key: &ApiKey,
+        api_key: ApiKey,
     ) -> Result<(), MintRequestError> {
         let timestamp = SystemTime::now()
             .duration_since(UNIX_EPOCH)
             .expect("current timestamp should be greater than unix epoch")
             .as_secs();
-        self.pow
-            .submit_challenge(account_id, api_key, challenge, nonce, timestamp)
+        let account_id_bytes: [u8; AccountId::SERIALIZED_SIZE] = account_id.into();
+        let mut requestor = [0u8; 32];
+        requestor[..AccountId::SERIALIZED_SIZE].copy_from_slice(&account_id_bytes);
+        self.rate_limiter
+            .submit_challenge(requestor, api_key, challenge, nonce, timestamp)
             .map_err(MintRequestError::PowError)
     }
 }
@@ -194,9 +200,23 @@ impl FromRef<Server> for GetTokensState {
     }
 }
 
-impl FromRef<Server> for PoW {
+impl FromRef<Server> for PoWRateLimiter {
     fn from_ref(input: &Server) -> Self {
         // Clone is cheap: only copies a 32-byte array and increments Arc reference counters.
-        input.pow.clone()
+        input.rate_limiter.clone()
     }
+}
+
+// ERRORS
+// ================================================================================================
+
+/// Errors that can occur when parsing an account ID or address.
+#[derive(Debug, thiserror::Error)]
+pub enum AccountError {
+    #[error("account ID failed to parse")]
+    ParseId(#[source] AccountIdError),
+    #[error("account address failed to parse")]
+    ParseAddress(#[source] AddressError),
+    #[error("account address is not an ID based")]
+    AddressNotIdBased,
 }