+++ bin/faucet/src/network.rs
@@ -7,8 +7,8 @@ use serde::{Deserialize, Serialize};
 // NETWORK
 // ================================================================================================
 
-/// Represents the network where the faucet is running. It is used to show the correct bech32
-/// addresses and explorer URL in the UI.
+/// Represents the network where the faucet is running. It is used to display the correct bech32
+/// addresses in the UI.
 #[derive(Debug, Clone, Deserialize, Serialize)]
 pub enum FaucetNetwork {
     Testnet,
@@ -39,20 +39,5 @@ impl FaucetNetwork {
             FaucetNetwork::Localhost => NetworkId::new("mlcl")?,
             FaucetNetwork::Custom(s) => NetworkId::new(s)?,
         })
-    }
-}
-
-/// A type wrapper for the explorer URL.
-#[derive(Debug, Clone, Serialize)]
-pub struct ExplorerUrl(&'static str);
-
-impl ExplorerUrl {
-    /// Returns the explorer URL for the given network ID.
-    /// Currently only testnet explorer is available.
-    pub fn from_network_id(network_id: NetworkId) -> Option<Self> {
-        match network_id {
-            NetworkId::Testnet => Some(Self("https://testnet.midenscan.com")),
-            _ => None,
-        }
     }
 }