+++ bin/faucet/src/testing/stub_rpc_api.rs
@@ -101,13 +101,6 @@ impl api_server::Api for StubRpcApi {
         unimplemented!()
     }
 
-    async fn get_account_state_delta(
-        &self,
-        _request: Request<proto::rpc_store::AccountStateDeltaRequest>,
-    ) -> Result<Response<proto::rpc_store::AccountStateDelta>, Status> {
-        unimplemented!()
-    }
-
     async fn get_account_proofs(
         &self,
         _request: Request<proto::rpc_store::AccountProofsRequest>,
@@ -119,6 +112,20 @@ impl api_server::Api for StubRpcApi {
         &self,
         _request: Request<()>,
     ) -> Result<Response<proto::rpc::RpcStatus>, Status> {
+        unimplemented!()
+    }
+
+    async fn sync_account_vault(
+        &self,
+        _request: Request<proto::rpc_store::SyncAccountVaultRequest>,
+    ) -> Result<Response<proto::rpc_store::SyncAccountVaultResponse>, Status> {
+        unimplemented!()
+    }
+
+    async fn sync_storage_maps(
+        &self,
+        _request: Request<proto::rpc_store::SyncStorageMapsRequest>,
+    ) -> Result<Response<proto::rpc_store::SyncStorageMapsResponse>, Status> {
         unimplemented!()
     }
 }